Address metadata get review feedback
