+++ tools/Azure.Mcp.Tools.Advisor/src/Commands/AdvisorJsonContext.cs
@@ -12,6 +12,12 @@ namespace Azure.Mcp.Tools.Advisor.Commands;
 [JsonSerializable(typeof(List<string>))]
 [JsonSerializable(typeof(RecommendationData))]
 [JsonSerializable(typeof(RecommendationMetadataData))]
+[JsonSerializable(typeof(RecommendationMetadataDataProperties))]
+[JsonSerializable(typeof(RecommendationMetadataResourceMetadata))]
+[JsonSerializable(typeof(RecommendationMetadataActionData))]
+[JsonSerializable(typeof(RecommendationMetadataSourceProperties))]
+[JsonSerializable(typeof(RecommendationMetadataServiceRetirementData))]
+[JsonSerializable(typeof(RecommendationMetadataServiceHealthData))]
 [JsonSerializable(typeof(Models.Recommendation))]
 [JsonSerializable(typeof(Models.RecommendationType))]
 [JsonSerializable(typeof(Models.RecommendationMetadata))]
