+++ tools/Azure.Mcp.Tools.Advisor/src/Services/Models/RecommendationMetadataData.cs
@@ -19,45 +19,3 @@ internal sealed record RecommendationMetadataData(RecommendationMetadataDataProp
     public static RecommendationMetadataData? FromJson(JsonElement source) =>
         JsonSerializer.Deserialize(source, AdvisorJsonContext.Default.RecommendationMetadataData);
 }
-
-internal sealed record RecommendationMetadataDataProperties(
-    string? RecommendationTypeId,
-    string? DisplayName,
-    string? Label,
-    string? RecommendationCategory,
-    string? RecommendationSubCategory,
-    string? RecommendationImpact,
-    double? PriorityScore,
-    string? PotentialBenefits,
-    string? DetailedDescription,
-    string? LearnMoreLink,
-    string? SupportedResourceType,
-    string? RecommendationScope,
-    string? RecommendationDataSourceQuery,
-    RecommendationMetadataResourceMetadata? ResourceMetadata,
-    List<RecommendationMetadataActionData>? Actions,
-    string? Language,
-    string? LastRefreshed,
-    RecommendationMetadataSourceProperties? SourceProperties);
-
-internal sealed record RecommendationMetadataResourceMetadata(
-    string? Singular,
-    string? Plural);
-
-internal sealed record RecommendationMetadataActionData(
-    string? ActionType,
-    string? Caption,
-    string? DocumentLink,
-    string? BladeName);
-
-internal sealed record RecommendationMetadataSourceProperties(
-    RecommendationMetadataServiceRetirementData? ServiceRetirement);
-
-internal sealed record RecommendationMetadataServiceRetirementData(
-    string? RetirementDate,
-    string? RetirementFeatureName,
-    RecommendationMetadataServiceHealthData? ServiceHealth);
-
-internal sealed record RecommendationMetadataServiceHealthData(
-    List<string>? TrackingIds,
-    List<string>? AshUrls);
