+++ tools/Azure.Mcp.Tools.Advisor/src/Services/Models/RecommendationMetadataDataProperties.cs
@@ -1,0 +1,24 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Azure.Mcp.Tools.Advisor.Services.Models;
+
+internal sealed record RecommendationMetadataDataProperties(
+    string? RecommendationTypeId,
+    string? DisplayName,
+    string? Label,
+    string? RecommendationCategory,
+    string? RecommendationSubCategory,
+    string? RecommendationImpact,
+    double? PriorityScore,
+    string? PotentialBenefits,
+    string? DetailedDescription,
+    string? LearnMoreLink,
+    string? SupportedResourceType,
+    string? RecommendationScope,
+    string? RecommendationDataSourceQuery,
+    RecommendationMetadataResourceMetadata? ResourceMetadata,
+    List<RecommendationMetadataActionData>? Actions,
+    string? Language,
+    string? LastRefreshed,
+    RecommendationMetadataSourceProperties? SourceProperties);
