+++ tools/Azure.Mcp.Tools.Advisor/src/Services/Models/RecommendationMetadataServiceHealthData.cs
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Azure.Mcp.Tools.Advisor.Services.Models;
+
+internal sealed record RecommendationMetadataServiceHealthData(
+    List<string>? TrackingIds,
+    List<string>? AshUrls);
