+++ tools/Azure.Mcp.Tools.Advisor/src/Services/Models/RecommendationMetadataServiceRetirementData.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Azure.Mcp.Tools.Advisor.Services.Models;
+
+internal sealed record RecommendationMetadataServiceRetirementData(
+    string? RetirementDate,
+    string? RetirementFeatureName,
+    RecommendationMetadataServiceHealthData? ServiceHealth);
