+++ tools/Azure.Mcp.Tools.Advisor/src/Services/Models/RecommendationMetadataSourceProperties.cs
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Azure.Mcp.Tools.Advisor.Services.Models;
+
+internal sealed record RecommendationMetadataSourceProperties(
+    RecommendationMetadataServiceRetirementData? ServiceRetirement);
